+++ packages/ui-utils/README.md
@@ -1,1 +1,3 @@
 # Vercel AI SDK - UI Framework Integration Implementation Utilities
+
+Important: this is an internal API. Expect breaking changes.
